lap_sensor/requirements.txt
django==2.0.7
djangorestframework==3.8.2
psycopg2==2.7.5
django-heroku==0.3.1
django-webpack-loader==0.6.0
django-rest-auth==0.9.3
djoser==1.2.0
djangorestframework-jwt==1.11.0
django-allauth== 0.36.0